+++ wsidicom/metadata/schema/json/image.py
@@ -25,7 +25,11 @@ from wsidicom.metadata.image import (
     LossyCompression,
 )
 from wsidicom.metadata.schema.common import LoadingSchema
-from wsidicom.metadata.schema.json.fields import PointMmJsonField, SizeMmJsonField
+from wsidicom.metadata.schema.json.fields import (
+    DateTimeOrDateJsonField,
+    PointMmJsonField,
+    SizeMmJsonField,
+)
 
 
 class ExtendedDepthOfFieldJsonSchema(LoadingSchema[ExtendedDepthOfField]):
@@ -71,6 +75,7 @@ class ImageJsonSchema(LoadingSchema[Image]):
     lossy_compressions = fields.List(
         fields.Nested(LossyCompressionJsonSchema()), allow_none=True
     )
+    content_datetime = DateTimeOrDateJsonField(allow_none=True)
 
     @property
     def load_type(self):